Слияние кода из svn mpc-hc, rev.4698, 4699, 4701;
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@260 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/apps/mplayerc/PPageInternalFilters.cpp
@@ -149,7 +149,7 @@ BOOL CPPageInternalFiltersListBox::OnToolTipNotify(UINT id, NMHDR* pNMHDR, LRESU
 {
 	TOOLTIPTEXT* pTTT = (TOOLTIPTEXT*)pNMHDR;
 
-	filter_t* f = (filter_t*)GetItemDataPtr(pNMHDR->idFrom);
+	filter_t* f = (filter_t*)GetItemDataPtr(static_cast<int>(pNMHDR->idFrom));
 	if (f->nHintID == 0) {
 		return FALSE;
 	}
